Fixed error in the PyFrensie Estimator.addResponseFunction()
The swig wrapping of the estimator class did not know how to properly handle a shared pointer to a response function. The problem was fixed by adding #include MonteCarlo_ResponseFunction,  %shared_ptr(MonteCarlo::ResponseFunction), %import MonteCarlo_ResponseFunction to MonteCarlo_Estimator.i.

A TEST_FOR_EXCEPTION to see if the reponse function was continuous has been commented out since this would not allow for delta response function in the adjoint case.

A bug in the PyFrensie wrapping of the ParticleState classes was fixed by adding the new ChargeParticleState class to the MonteCarlo_ParticleState.i file.