Update README.md
This PR is intended to revise some of the language of the instructions to be more concept based, as opposed to just providing commands. Also, some updates to links/versions/commands that I noticed didn't work during my most recent build. Need to do another sweep on a different day to re-read the changes with fresh eyes.